Move humanfriendly_time_parser() into common
